pyproject.toml
[project]
name = "data-mirage-api"
version = "2.0.0"
description = "Data Mirage - Advanced 3D data visualization with ML processing API"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and ASGI
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.12.1",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # ML and Data Processing
    "scikit-learn>=1.3.2",
    "pandas>=2.1.3",
    "numpy>=1.24.4",
    "umap-learn>=0.5.4",
    "hdbscan>=0.8.33",
    
    # AI and RAG
    "openai>=1.3.0",
    
    # Redis for caching and sessions
    "redis>=5.0.1",
    
    # File processing
    "openpyxl>=3.1.2",
    "xlrd>=2.0.1",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    
    # Rate limiting
    "slowapi>=0.1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "umap.*",
    "hdbscan.*",
    "sklearn.*",
    "pandas.*",
    "numpy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
